pyproject.toml
[project]
name = "metagit"
version = "0.1.0"
description = "Metagit-ai multi-repo management tool"
readme = "README.md"
license-files = ["LICENSE"]

authors = [
    {name = "Zachary Loeber", email = "zloeber@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "logging>=0.4.9.6",
    "loguru>=0.7.3",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "click>=8.1.7",
    "dotenv>=0.9.9",
    "gitpython>=3.1.44",
    "rich>=14.0.0",
    "pyyaml>=6.0.2",
    "pygithub>=2.6.1",
    "jsonschema>=4.24.0",
    "python-dotenv>=1.1.0",
    "rapidfuzz>=3.13.0",
    "prompt-toolkit>=3.0.51",
    "tqdm>=4.66.4",
    "types-tqdm>=4.67.0.20250516",
    "aiofiles>=23.2.1",
    "async>=0.6.2",
]

[project.urls]
Homepage = "https://github.com/metagit-ai/metagit-detect"
Documentation = "https://metagit-ai.github.io/metagit-detect/"
Issues = "https://github.com/metagit-ai/metagit-detect/issues"
CI = "https://github.com/metagit-ai/metagit-detect/actions"

[project.optional-dependencies]
standard = ["crewai[tools]"]
test = ["pytest", "pytest-cov", "mypy", "ruff", "httpx", "pytest-asyncio", "black"]
dev = ["ruff", "black", "isort", "mypy"]
docs = ["mkdocs", "mkdocs-material", "mkdocstrings", "mkdocs-click", "mdtoc"]
lint = ["ruff", "black", "isort", "mypy"]
format = ["black", "isort"]
typecheck = ["mypy"]

[project.scripts]
metagit = "metagit.cli.main:cli"

[lint]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293', 'E501', 'SIM115']

[tool.ruff]
exclude = [".venv", "records", ".metagit", "examples", "docs", "tests"]

[dependency-groups]
dev = [
    "diagrams>=0.24.4",
    "mypy>=1.16.1",
    "ruff>=0.11.11",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["metagit", "metagit.*"]
exclude = ["tests", "tests.*"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

